travaia-ai/backend/application-job-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.24.0
google-cloud-firestore==2.13.1
google-cloud-pubsub==2.18.4
firebase-admin==6.2.0
pydantic[email]==2.5.0
structlog==23.2.0
python-multipart==0.0.6
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0
slowapi==0.1.9
aiohttp==3.9.1
tenacity==8.2.3
PyJWT==2.8.0